requirements.txt
Django>=4.0,<4.1
wagtail>=3.0
django-simple-history==3.1.1
django-simple-captcha==0.5.17
django-allauth==0.51.0